in-progress BuildMaster work
reworked the tool settings to be more generic.
added "entries" which are jobs (maybe they should be called jobs?) that use tools. can specify multiple entries for the same tool.
updated the config.cm file to use the new style
BuildMaster exports a list of directories to a makefile.config file, which will be used by the makefile later.
animations are now exported to exported/animations
maps are now exported to exported/maps
the game doesn't build with those folders yet