Archiving SpriteMaster C#
Building a new converter from scratch in C# that includes a WPF interface requires too much effort for the amount I want to give it. I'll give some love to gg2c instead.
Plus, having a .sm file for each .gal was going to be a pain in the ass to synchronize and validate. It wasn't workable or friendly to work with. Storing the data inside the .gal is best.